backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
openai
python-dotenv
sse-starlette
pdfplumber
